# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Learn Hub",
        "description": "API doc for Learn Hub",
        "contact": {
            "url": "https://justice-chimobi.vercel.app/"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://learn-hub-backend.test/api/v1"
        }
    ],
    "paths": {
        "/articles/{article}/comments": {
            "post": {
                "tags": [
                    "articles"
                ],
                "summary": "Write a comment on a specific articles",
                "description": "Write a comment on a specific articles",
                "operationId": "1b06e3f945e94c27c58e12cf601a2900",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "string"
                        },
                        "examples": {
                            "ID": {
                                "summary": "An UUID value.",
                                "value": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                            }
                        }
                    }
                ],
                "requestBody": {
                    "description": "Write a comment on a specific articles",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "comment": {
                                        "type": "string",
                                        "example": "Hey i love your content and keep it up"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Comment successfully"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
        "/articles/{article}": {
            "get": {
                "tags": [
                    "articles"
                ],
                "summary": "Get an article by ID",
                "description": "Get an article by ID",
                "operationId": "fd2268af73be3313792b6fc4943bed84",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Article ID",
                        "required": true,
                        "schema": {
                            "type": "number"
                        },
                        "examples": {
                            "id": {
                                "summary": "An ID value.",
                                "value": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/ArticleResource"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "404": {
                        "description": "Article not found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/auth/login": {
            "post": {
                "tags": [
                    "auth"
                ],
                "summary": "Logs an existing user with credentials",
                "description": "authentication endpoint using JWT token",
                "operationId": "fbb752a9447382cb9e22985c9b58276e",
                "requestBody": {
                    "description": "authenticate if user exists",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "justice@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "secrets"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Logged in successfully",
                                    "access_token": "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJpc3MiOiJodHRwOi8vbG9jYWxob3N0OjgwMDAvYXBpL2F1dGgvc2lnbAsIm5iZiI6MTcwNjAwNDI1MCwianRpIjoidE1tUlNHbjg0QlZoQkdYTSIsInN1YiI6IjU1MjJmNjQyLTYyOGQtNDZkZi04NDMyLTNmZTBjMjllNDM5YiIsInBydiI6IjIzYmQ1Yzg5NDlmNjAwYWRiMzllNzAxYzQwMDg3MmRiN2E1OTc2ZjcifQ.xoKfHEHLZ6vDwIfg-GZQXoMwG9wQGlgs4vDrLTEVwoU"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "422": {
                        "description": "Unprocessable Content"
                    }
                }
            }
        },
...
        "/image/upload": {
            "post": {
                "tags": [
                    "upload"
                ],
                "summary": "Upload image to cloudinary",
                "description": "Upload image to cloudinary and get actual url from response to use for avatar, thumbnail endpoint",
                "operationId": "42f886e2e4b61acd6302b69d5a6417f7",
                "requestBody": {
                    "description": "Image file to upload",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "image": {
                                        "description": "image file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Uploaded successfully",
                        "content": {
                            "application/json": {
                                "schema": {},
                                "example": {
                                    "message": "Uploaded successfully!",
                                    "data": {
                                        "imageUploadUrl": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.png"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "ArticleResource": {
                "title": "ArticleResource",
                "description": "Article Resource",
                "properties": {
                    "id": {
                        "description": "Course ID",
                        "type": "string",
                        "example": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                    },
                    "title": {
                        "description": "Article title",
                        "type": "string",
                        "example": "Article intro"
                    },
                    "slug": {
                        "description": "Article slug title",
                        "type": "string",
                        "example": "article-intro"
                    },
                    "content": {
                        "description": "Article content",
                        "type": "string",
                        "example": "About my article content."
                    },
                    "thumbnail": {
                        "description": "URL of the article thumbnail",
                        "type": "string",
                        "example": "https://res.cloudinary.com/dbx3dhfkt/image/upload/v1672045944/estudy/pictures/image-5a9482cd3-a97e-4627-dbc3-9cb53797e40a.png"
                    },
                    "read_time": {
                        "description": "Article content time read",
                        "type": "string",
                        "example": "25 mins read"
                    },
                    "created_at": {
                        "$ref": "#/components/schemas/DateTimeResource"
                    },
                    "author": {
                        "$ref": "#/components/schemas/UserShortResource"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "ArticleResource"
                }
            },
...
            "DateTimeResource": {
                "title": "DateTimeResource",
                "description": "format date time",
                "properties": {
                    "human": {
                        "description": "long format string",
                        "type": "string",
                        "example": "2 hours ago"
                    },
                    "date_time": {
                        "description": "string representation",
                        "type": "string",
                        "example": "2024-01-22 12:01:58"
                    },
                    "human_short": {
                        "description": "short format string",
                        "type": "string",
                        "example": "2h ago"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "DateTimeResource"
                }
            },
...
            "UserShortResource": {
                "title": "UserShortResource",
                "description": "User article Resource",
                "properties": {
                    "id": {
                        "description": "Id",
                        "type": "string",
                        "example": "9d04e3b2-7e14-472e-8095-7f7f9f0c943f"
                    },
                    "fullname": {
                        "description": "Full name",
                        "type": "string",
                        "example": "Justice Chimobi"
                    },
                    "username": {
                        "description": "Username",
                        "type": "string",
                        "example": "justice-chimobi"
                    },
                    "avatar": {
                        "description": "Course name",
                        "type": "string",
                        "example": "https://res.cloudinary.com/estudy/image/upload/v1705789451/yofikr4gyecw04sp5ial.jpg"
                    },
                    "profile_headlines": {
                        "description": "profile_headlines",
                        "type": "string",
                        "example": "Frontend Developer || Full Stack"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "UserShortResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
...
}